Render toast if the pricing service is down
