+++ public/ibc_assets.json
@@ -24,7 +24,8 @@
       "channel": "info",
       "juno_channel": "channel-7",
       "juno_denom": "ibc/test",
-      "logoURI": "https://cryptologos.cc/logos/usd-coin-usdc-logo.svg?v=014"
+      "logoURI": "https://cryptologos.cc/logos/usd-coin-usdc-logo.svg?v=014",
+      "external_deposit_uri": ""
     }
   ]
 }
